replace all background tasks to celery tasks
