+++ chatbot/views.py
@@ -1,7 +1,3 @@
-import socket
-from threading import Thread
-
-import requests
 from django.contrib.auth import authenticate, password_validation
 from django.core import exceptions
 from django.db.utils import IntegrityError
@@ -16,16 +12,7 @@ from rest_framework.status import HTTP_400_BAD_REQUEST, HTTP_404_NOT_FOUND, HTTP
 
 from chatbot.models import Muser, Message
 from chatbot.serializers import MuserSerializer, MessageSerializer
-
-
-def chatscript(username, text):
-    sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-    sock.connect(('localhost', 1024))
-    content = f'{username}\0\0{text}\0'
-    sock.send(bytes(content, encoding='utf-8'))
-    response = sock.recv(1024).decode('utf-8')
-    sock.close()
-    return response
+from chatbot.tasks import respond, chatscript, greet
 
 
 @api_view(['POST'])
@@ -49,11 +36,6 @@ def signup(request):
         return Response({'error': e}, status=HTTP_400_BAD_REQUEST)
 
 
-def greet(user):
-    text = chatscript(user.username, '')
-    serializer = MessageSerializer(data=dict(receiver=user, text=text))
-    if serializer.is_valid():
-        serializer.save()
 
 
 @api_view(['POST'])
@@ -76,7 +58,7 @@ def signin(request):
     device.active = True
     device.save()
 
-    Thread(target=greet, args=(user,)).start()
+    greet.delay(user.id)
 
     token, _ = Token.objects.get_or_create(user=user)
     return Response({'token': token.key}, status=HTTP_200_OK)
@@ -159,30 +141,11 @@ class Chat(generics.ListCreateAPIView):
         user = self.request.user
         return Message.objects.filter(sender=user) | Message.objects.filter(receiver=user)
 
-    def respond(self, user, text):
-        from random import sample
-
-        text = chatscript(user.username, text)
-        chips = sample(range(0, 20), 4)
-        message = Message.objects.create(receiver=user, text=text, chips=chips)
-
-        device = FCMDevice.objects.filter(user=user).first()
-        if device:
-            retry = 10
-            for _ in range(retry):
-                try:
-                    device.send_message(data={'message_id': message.id, 'text': text})
-                except requests.exceptions.ReadTimeout as e:
-                    print(e)
-                else:
-                    break
-
     def perform_create(self, serializer):
         user = Muser.objects.get_by_natural_key(self.request.user.username)
         text = self.request.data.get('text')
-
-        Thread(target=self.respond, args=(user, text)).start()
         serializer.save(sender=user, text=text)
+        respond.delay(user.id, text)
 
 
 class ChatDetail(generics.RetrieveAPIView):
