+++ manager/apps.py
@@ -3,3 +3,6 @@ from django.apps import AppConfig
 
 class ManagerConfig(AppConfig):
     name = 'manager'
+
+    def ready(self):
+        import manager.tasks
