+++ requirements.txt
@@ -1,4 +1,7 @@
+amqp==2.3.2
 beautifulsoup4==4.6.3
+billiard==3.5.0.4
+celery==4.2.1
 certifi==2018.10.15
 chardet==3.0.4
 Django==2.1.2
@@ -7,9 +10,11 @@ djangorestframework==3.8.2
 fcm-django==0.2.19
 future==0.17.1
 idna==2.7
+kombu==4.2.1
 pyfcm==1.4.5
 pytz==2018.5
 requests==2.20.0
 requests-toolbelt==0.8.0
 serializers==0.2.4
 urllib3==1.24
+vine==1.1.4
